+++ example/lib/page/style/bezier_page.dart
@@ -239,6 +239,8 @@ class _BezierPageState extends State<BezierPage> {
           foregroundColor: themeData.colorScheme.onPrimary,
           clamping: _headerProperties.clamping,
           showBalls: _headerProperties.displayBalls,
+          spinInCenter: _headerProperties.spinInCenter,
+          onlySpin: _headerProperties.onlySpin,
           spinWidget: _spinWidget(
               _headerProperties.spin, themeData.colorScheme.onPrimary),
         ),
@@ -246,7 +248,9 @@ class _BezierPageState extends State<BezierPage> {
           backgroundColor: themeData.colorScheme.primary,
           foregroundColor: themeData.colorScheme.onPrimary,
           clamping: _footerProperties.clamping,
-          showBalls: _headerProperties.displayBalls,
+          showBalls: _footerProperties.displayBalls,
+          spinInCenter: _footerProperties.spinInCenter,
+          onlySpin: _footerProperties.onlySpin,
           spinWidget: _spinWidget(
               _headerProperties.spin, themeData.colorScheme.onPrimary),
         ),
@@ -394,6 +398,28 @@ class _BezierPageState extends State<BezierPage> {
                                 },
                               ),
                             ),
+                            ListTile(
+                              title: Text('Spin in center'.tr),
+                              trailing: Switch(
+                                value: properties.spinInCenter,
+                                onChanged: (value) {
+                                  setState(() {
+                                    properties.spinInCenter = value;
+                                  });
+                                },
+                              ),
+                            ),
+                            ListTile(
+                              title: Text('Only show spin'.tr),
+                              trailing: Switch(
+                                value: properties.onlySpin,
+                                onChanged: (value) {
+                                  setState(() {
+                                    properties.onlySpin = value;
+                                  });
+                                },
+                              ),
+                            ),
                             const ListTile(
                               title: Text('Spin'),
                             ),
@@ -441,6 +467,8 @@ class _BIProperties {
   bool clamping = false;
   bool displayBalls = true;
   String spin = 'HourGlass';
+  bool spinInCenter = true;
+  bool onlySpin = false;
 
   _BIProperties({
     required this.name,