+++ lib/src/easy_paging.dart
@@ -32,10 +32,10 @@ abstract class EasyPaging<DataType, ItemType> extends StatefulWidget {
   final bool simultaneously;
 
   /// Is it possible to refresh after there is no more.
-  final bool noMoreRefresh;
+  final bool canRefreshAfterNoMore;
 
   /// Is it loadable after no more.
-  final bool noMoreLoad;
+  final bool canLoadAfterNoMore;
 
   /// Reset after refresh when no more deactivation is loaded.
   final bool resetAfterRefresh;
@@ -76,8 +76,8 @@ abstract class EasyPaging<DataType, ItemType> extends StatefulWidget {
     this.notRefreshHeader,
     this.notLoadFooter,
     this.simultaneously = false,
-    this.noMoreRefresh = false,
-    this.noMoreLoad = false,
+    this.canRefreshAfterNoMore = false,
+    this.canLoadAfterNoMore = false,
     this.resetAfterRefresh = true,
     this.refreshOnStart = false,
     this.callRefreshOverOffset = 20,
@@ -141,6 +141,13 @@ abstract class EasyPagingState<DataType, ItemType> extends State<EasyPaging> {
 
   Future _onRefresh() async {
     final result = await onRefresh();
+    if (!_refreshController.controlFinishRefresh && isNoMore) {
+      _ambiguate(WidgetsBinding.instance)!.addPostFrameCallback((timeStamp) {
+        if (mounted) {
+          _refreshController.finishLoad(IndicatorResult.noMore, true);
+        }
+      });
+    }
     if (result is IndicatorResult) {
       return result;
     }
@@ -159,6 +166,23 @@ abstract class EasyPagingState<DataType, ItemType> extends State<EasyPaging> {
       return IndicatorResult.noMore;
     }
     return IndicatorResult.success;
+  }
+
+  /// EasyRefresh controller.
+  late EasyRefreshController _refreshController;
+
+  @override
+  void initState() {
+    super.initState();
+    _refreshController = widget.controller ?? EasyRefreshController();
+  }
+
+  @override
+  void didUpdateWidget(covariant EasyPaging oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.controller != widget.controller) {
+      _refreshController = widget.controller ?? EasyRefreshController();
+    }
   }
 
   /// Empty widget.
@@ -270,8 +294,8 @@ abstract class EasyPagingState<DataType, ItemType> extends State<EasyPaging> {
         notRefreshHeader: widget.notRefreshHeader,
         notLoadFooter: widget.notLoadFooter,
         simultaneously: widget.simultaneously,
-        noMoreRefresh: widget.noMoreRefresh,
-        noMoreLoad: widget.noMoreLoad,
+        canRefreshAfterNoMore: widget.canRefreshAfterNoMore,
+        canLoadAfterNoMore: widget.canLoadAfterNoMore,
         resetAfterRefresh: widget.resetAfterRefresh,
         refreshOnStart: widget.refreshOnStart,
         callRefreshOverOffset: widget.callRefreshOverOffset,
@@ -287,14 +311,14 @@ abstract class EasyPagingState<DataType, ItemType> extends State<EasyPaging> {
       refreshOnStartHeader: startHeader,
       onRefresh: _onRefresh,
       onLoad: _onLoad,
-      controller: widget.controller,
+      controller: _refreshController,
       spring: widget.spring,
       frictionFactor: widget.frictionFactor,
       notRefreshHeader: widget.notRefreshHeader,
       notLoadFooter: widget.notLoadFooter,
       simultaneously: widget.simultaneously,
-      noMoreRefresh: widget.noMoreRefresh,
-      noMoreLoad: widget.noMoreLoad,
+      canRefreshAfterNoMore: widget.canRefreshAfterNoMore,
+      canLoadAfterNoMore: widget.canLoadAfterNoMore,
       resetAfterRefresh: widget.resetAfterRefresh,
       refreshOnStart: widget.refreshOnStart,
       callRefreshOverOffset: widget.callRefreshOverOffset,