+++ lib/src/styles/bezier/bezier_indicator.dart
@@ -37,12 +37,21 @@ class _BezierIndicator extends StatefulWidget {
   /// Background color.
   final Color? backgroundColor;
 
+  /// Whether the spin widget is in the center.
+  final bool spinInCenter;
+
+  /// Only display the spin.
+  /// When true, the balls are no longer displayed.
+  final bool onlySpin;
+
   const _BezierIndicator({
     Key? key,
     required this.state,
     required this.reverse,
     required this.processedDuration,
     this.showBalls = true,
+    this.spinInCenter = true,
+    this.onlySpin = false,
     this.spinWidget,
     this.noMoreWidget,
     this.spinBuilder,
@@ -69,6 +78,8 @@ class _BezierIndicatorState extends State<_BezierIndicator>
   double get _actualTriggerOffset => widget.state.actualTriggerOffset;
 
   double get _safeOffset => widget.state.safeOffset;
+
+  bool get _reverse => widget.reverse;
 
   /// Animation controller.
   late AnimationController _animationController;
@@ -103,24 +114,134 @@ class _BezierIndicatorState extends State<_BezierIndicator>
     }
   }
 
+  /// Build balls.
+  Widget _buildBalls() {
+    final extent = (!widget.spinInCenter && _offset > _actualTriggerOffset)
+        ? _actualTriggerOffset
+        : null;
+    final full = widget.spinInCenter ||
+        (!widget.spinInCenter && _offset <= _actualTriggerOffset);
+    return Positioned(
+      top: (_axis == Axis.vertical && _reverse && !full) ? null : 0,
+      bottom: (_axis == Axis.vertical && !_reverse && !full) ? null : 0,
+      left: (_axis == Axis.horizontal && _reverse && !full) ? null : 0,
+      right: (_axis == Axis.horizontal && !_reverse && !full) ? null : 0,
+      height: _axis == Axis.vertical ? extent : null,
+      width: _axis == Axis.horizontal ? extent : null,
+      child: LayoutBuilder(
+        builder: (ctx, constraints) {
+          final ballSize = math.min(_ballSize, _offset);
+          final scale = _offset / _actualTriggerOffset;
+          final length = _axis == Axis.vertical
+              ? constraints.maxWidth
+              : constraints.maxHeight;
+          final ballArea = math.min(_ballArea, length);
+          return Stack(
+            alignment: Alignment.center,
+            children: [
+              for (int i = 0; i < 7; i++)
+                Positioned(
+                  left: _axis == Axis.vertical
+                      ? ((length - ballSize) / 2) +
+                          (ballArea / 8 * scale) * (i - 3)
+                      : null,
+                  bottom: _axis == Axis.horizontal
+                      ? ((length - ballSize) / 2) +
+                          (ballArea / 8 * scale) * (i - 3)
+                      : null,
+                  child: AnimatedBuilder(
+                    animation: _animationController,
+                    builder: (ctx, _) {
+                      double aValue;
+                      if (_mode == IndicatorMode.inactive ||
+                          _mode == IndicatorMode.drag ||
+                          _mode == IndicatorMode.armed) {
+                        aValue = 1;
+                      } else {
+                        aValue = _animationController.isAnimating
+                            ? 1 - _animationController.value
+                            : 0;
+                      }
+                      return Opacity(
+                        opacity: (1 - ((0.8 / 3) * (i - 3).abs())) *
+                            math.min(1, scale) *
+                            aValue,
+                        child: Container(
+                          width: ballSize,
+                          height: ballSize,
+                          decoration: BoxDecoration(
+                            color: _foregroundColor,
+                            borderRadius:
+                                BorderRadius.all(Radius.circular(ballSize / 2)),
+                          ),
+                        ),
+                      );
+                    },
+                  ),
+                ),
+            ],
+          );
+        },
+      ),
+    );
+  }
+
   /// Build spin widget.
   Widget _buildSpin() {
+    Widget spinWidget;
     if (widget.spinBuilder != null) {
-      widget.spinBuilder!(context, _animationController.value);
+      spinWidget = widget.spinBuilder!(context, _animationController.value);
+    } else {
+      spinWidget = widget.spinWidget ??
+          _SpinKitHourGlass(
+            color: _foregroundColor,
+            size: 32,
+          );
     }
-    return AnimatedBuilder(
+    Widget animatedWidget = AnimatedBuilder(
       animation: _animationController,
       builder: (context, _) {
         return Transform.scale(
-          scale: _animationController.value,
-          child: widget.spinWidget ??
-              _SpinKitHourGlass(
-                color: _foregroundColor,
-                size: 32,
-              ),
+          scale: widget.onlySpin ? 1 : _animationController.value,
+          child: spinWidget,
         );
       },
     );
+    if (widget.onlySpin && _offset < _actualTriggerOffset) {
+      return Positioned(
+        top: (_axis == Axis.vertical && !_reverse)
+            ? -(_actualTriggerOffset - _offset) / 2
+            : null,
+        bottom: (_axis == Axis.vertical && _reverse)
+            ? -(_actualTriggerOffset - _offset) / 2
+            : null,
+        left: (_axis == Axis.horizontal && !_reverse)
+            ? -(_actualTriggerOffset - _offset) / 2
+            : null,
+        right: (_axis == Axis.horizontal && _reverse)
+            ? -(_actualTriggerOffset - _offset) / 2
+            : null,
+        height: _axis == Axis.vertical ? _actualTriggerOffset : null,
+        width: _axis == Axis.vertical ? null : _actualTriggerOffset,
+        child: Center(
+          child: animatedWidget,
+        ),
+      );
+    }
+    if (!widget.spinInCenter) {
+      return Positioned(
+        top: (_axis == Axis.vertical && _reverse) ? null : 0,
+        bottom: (_axis == Axis.vertical && !_reverse) ? null : 0,
+        left: (_axis == Axis.horizontal && _reverse) ? null : 0,
+        right: (_axis == Axis.horizontal && !_reverse) ? null : 0,
+        height: _axis == Axis.vertical ? _actualTriggerOffset : null,
+        width: _axis == Axis.vertical ? null : _actualTriggerOffset,
+        child: Center(
+          child: animatedWidget,
+        ),
+      );
+    }
+    return animatedWidget;
   }
 
   @override
@@ -141,70 +262,14 @@ class _BezierIndicatorState extends State<_BezierIndicator>
           color: widget.backgroundColor,
         ),
         if (widget.showBalls &&
+            !widget.onlySpin &&
             _offset > _safeOffset &&
             widget.state.result != IndicatorResult.noMore)
-          Positioned(
-            top: 0,
-            bottom: 0,
-            left: 0,
-            right: 0,
-            child: LayoutBuilder(
-              builder: (ctx, constraints) {
-                final ballSize = math.min(_ballSize, _offset);
-                final scale = _offset / _actualTriggerOffset;
-                final length = _axis == Axis.vertical
-                    ? constraints.maxWidth
-                    : constraints.maxHeight;
-                final ballArea = math.min(_ballArea, length);
-                return Stack(
-                  alignment: Alignment.center,
-                  children: [
-                    for (int i = 0; i < 7; i++)
-                      Positioned(
-                        left: _axis == Axis.vertical
-                            ? ((length - ballSize) / 2) +
-                                (ballArea / 8 * scale) * (i - 3)
-                            : null,
-                        bottom: _axis == Axis.horizontal
-                            ? ((length - ballSize) / 2) +
-                                (ballArea / 8 * scale) * (i - 3)
-                            : null,
-                        child: AnimatedBuilder(
-                          animation: _animationController,
-                          builder: (ctx, _) {
-                            double aValue;
-                            if (_mode == IndicatorMode.inactive ||
-                                _mode == IndicatorMode.drag ||
-                                _mode == IndicatorMode.armed) {
-                              aValue = 1;
-                            } else {
-                              aValue = _animationController.isAnimating
-                                  ? 1 - _animationController.value
-                                  : 0;
-                            }
-                            return Opacity(
-                              opacity: (1 - ((0.8 / 3) * (i - 3).abs())) *
-                                  math.min(1, scale) *
-                                  aValue,
-                              child: Container(
-                                width: ballSize,
-                                height: ballSize,
-                                decoration: BoxDecoration(
-                                  color: _foregroundColor,
-                                  borderRadius: BorderRadius.all(
-                                      Radius.circular(ballSize / 2)),
-                                ),
-                              ),
-                            );
-                          },
-                        ),
-                      ),
-                  ],
-                );
-              },
-            ),
-          ),
-        if (_mode == IndicatorMode.ready || _mode == IndicatorMode.processing)
+          _buildBalls(),
+        if (_mode == IndicatorMode.ready ||
+            _mode == IndicatorMode.processing ||
+            (widget.onlySpin &&
+                (_mode == IndicatorMode.drag || _mode == IndicatorMode.armed)))
           _buildSpin(),
         if (widget.state.result == IndicatorResult.noMore)
           Positioned(